Split RemotePlugin.h into separate files (#6322)
