+++ plugins/zynaddsubfx/RemoteZynAddSubFx.cpp
@@ -31,9 +31,7 @@
 
 #undef CursorShape // is, by mistake, not undefed in FL
 
-#define BUILD_REMOTE_PLUGIN_CLIENT
-
-#include "RemotePlugin.h"
+#include "RemotePluginClient.h"
 #include "RemoteZynAddSubFx.h"
 #include "LocalZynAddSubFx.h"
