+++ src/CMakeLists.txt
@@ -16,9 +16,12 @@ IF(LMMS_BUILD_APPLE)
 	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -stdlib=libc++")
 ENDIF()
 
+ADD_SUBDIRECTORY(common)
 ADD_SUBDIRECTORY(core)
 ADD_SUBDIRECTORY(gui)
 ADD_SUBDIRECTORY(tracks)
+
+LIST(APPEND LMMS_SRCS ${LMMS_COMMON_SRCS})
 
 QT5_WRAP_UI(LMMS_UI_OUT ${LMMS_UIS})
 INCLUDE_DIRECTORIES(
