+++ src/common/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(COMMON_SRCS
+	RemotePluginBase.cpp
+)
+
+foreach(SRC ${COMMON_SRCS})
+	list(APPEND LMMS_COMMON_SRCS "${CMAKE_CURRENT_SOURCE_DIR}/${SRC}")
+endforeach()
+
+set(LMMS_COMMON_SRCS ${LMMS_COMMON_SRCS} PARENT_SCOPE)
