+++ src/core/RemotePlugin.cpp
@@ -22,17 +22,18 @@
  *
  */
 
-#define COMPILE_REMOTE_PLUGIN_BASE
+#include "RemotePlugin.h"
+
 //#define DEBUG_REMOTE_PLUGIN
 #ifdef DEBUG_REMOTE_PLUGIN
 #include <QDebug>
 #endif
 
 #include "BufferManager.h"
-#include "RemotePlugin.h"
 #include "AudioEngine.h"
 #include "Engine.h"
 
+#include <QCoreApplication>
 #include <QDebug>
 #include <QDir>
